Move error logs in common and download errorlog when using -o.
