+++ packages/common/src/index.ts
@@ -24,6 +24,7 @@ export {
 } from './utils/findAnalyticsMatch';
 export { default as removeAndAddNewSpinnerText } from './utils/removeAndAddNewSpinnerText';
 export { default as calculateEffectiveExpiryDate } from './utils/calculateEffectiveExpiryDate';
+export { default as generateErrorLogFile } from './utils/generateErrorLogs';
 export { default as sanitizeCsvRecord } from './utils/sanitizeCsvRecord';
 export { parseUrl } from './utils/parseUrl';
 export { default as fetchLocalData } from './utils/fetchLocalData';
